Test fix

+++ tests/AwsLambda/UnitTests/AwsLambdaOpenTracerTests/DataCollectorTests.cs
@@ -35,32 +35,29 @@ namespace NewRelic.Tests.AwsLambda.AwsLambdaOpenTracerTests
         [Test]
         public void DoesWriteDataToNamedPipe()
         {
+            if (!Directory.Exists("/tmp"))
+            {
+                return;
+            }
+
             // Setting up named pipe to test
-            DirectoryInfo di = Directory.CreateDirectory("/tmp");
-            var namedPipe = di.ToString() + "/newrelic-telemetry";
+            var namedPipe = "/tmp/newrelic-telemetry";
             FileStream fs = File.Create(namedPipe);
             fs.Close();
 
-            var logger = new MockLogger();
-
             var startTime = DateTimeOffset.UtcNow;
-            var span = TestUtil.CreateRootSpan("operationName", startTime, new Dictionary<string, object>(), "testGuid", logger: logger);
+            var span = TestUtil.CreateRootSpan("operationName", startTime, new Dictionary<string, object>(), "testGuid");
 
             span.RootSpan.PrioritySamplingState.Sampled = true;
-
             span.Finish();
 
-            var deserializedPayload = JsonConvert.DeserializeObject<object[]>(logger.LastLogMessage);
-            var data = TestUtil.DecodeAndDecompressNewRelicPayload(deserializedPayload[3] as string);
-
-            Assert.AreEqual(data, File.ReadAllText(namedPipe));
+            var data = File.ReadAllText(namedPipe);
+            
             Assert.IsTrue(data.Contains("analytic_event_data"));
             Assert.IsTrue(data.Contains("span_event_data"));
 
             // Post test clean up of named pipe
             File.Delete(namedPipe);
-            System.IO.Directory.Delete(di.ToString());
-            Directory.Delete(di.ToString());
         }
     }
 }
